backend/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-extensions==3.1.3
djangorestframework==3.12.4
environ==1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
model_mommy==2.0.0

